finished phase 2

+++ treasure_hub.c
@@ -11,6 +11,7 @@
 #define MAX_USERNAME_LEN 32
 #define MAX_CLUE_LEN 256
 #define MAX_HUNT_ID_LEN 128
+#define MAX_TREASURE_ID_LEN 32
 
 typedef struct {
     int id;
@@ -22,8 +23,6 @@ typedef struct {
 } Treasure;
 
 pid_t monitor_pid = 0;
-
-char cmd_hunt_id[MAX_HUNT_ID_LEN];
 
 void exec_list(char *hunt_id)
 {
@@ -70,7 +69,104 @@ void list_hunts(int sig)
 
 void list_treasures(int sig)
 {
-    exec_list(cmd_hunt_id);
+    char temp_filename[64];
+    snprintf(temp_filename, sizeof(temp_filename), "list_treasures_%d", getppid());
+
+    //printf("%s\n", temp_filename);
+    int fd = open(temp_filename, O_RDONLY);
+    if(fd == -1)
+    {
+        perror("(list_treasures) open");
+        exit(1);
+    }
+
+    char hunt_id[MAX_HUNT_ID_LEN];
+    if(read(fd, hunt_id, MAX_HUNT_ID_LEN - 1) > 0)
+    {
+        hunt_id[strlen(hunt_id)] = '\0';
+        exec_list(hunt_id);
+    }
+    else
+    {
+        perror("(list_treasures) read");
+        close(fd);
+        unlink(temp_filename);
+        exit(1);
+    }
+    close(fd);
+    unlink(temp_filename); //deleting the temp file
+}
+
+void exec_view(char *hunt_id, char *id)
+{
+    pid_t view_pid = 0;
+    if((view_pid = fork()) == -1)
+    {
+        perror("(exec_view) fork");
+        exit(1);
+    }
+    if(view_pid == 0)
+    {
+        execlp("./treasure_manager", "./treasure_manager", "--view", hunt_id, id, NULL);
+        //the program shouldn't reach this part only if exec failed
+        perror("(exec_view) execlp:");
+        exit(1);
+    }
+    int status;
+    waitpid(view_pid, &status, 0);
+}
+
+void view_treasure(int sig)
+{
+    char hunt_id[MAX_HUNT_ID_LEN];
+    char id[MAX_TREASURE_ID_LEN];
+
+    char temp_filename[64];
+    snprintf(temp_filename, sizeof(temp_filename), "view_treasure_hunt_id_%d", getppid());
+    int fd = open(temp_filename, O_RDONLY);
+    if(fd == -1)
+    {
+        perror("(view_treasure) open");
+        exit(1);
+    }
+    if(read(fd, hunt_id, MAX_HUNT_ID_LEN - 1) > 0)
+    {
+        hunt_id[strlen(hunt_id)] = '\0';
+    }
+    else
+    {
+        perror("(list_treasures) read");
+        close(fd);
+        unlink(temp_filename);
+        exit(1);
+    }
+    close(fd);
+    unlink(temp_filename); //deleting the temp file
+
+    char temp_filename1[64];
+    snprintf(temp_filename1, sizeof(temp_filename1), "view_treasure_id_%d", getppid());
+    int fd1 = open(temp_filename1, O_RDONLY);
+    if(fd1 == -1)
+    {
+        perror("(view_treasure) open");
+        exit(1);
+    }
+    if(read(fd1, id, MAX_HUNT_ID_LEN - 1) > 0)
+    {
+        id[strlen(id)] = '\0';
+    }
+    else
+    {
+        perror("(list_treasures) read");
+        close(fd1);
+        unlink(temp_filename1);
+        exit(1);
+    }
+
+    close(fd1);
+    unlink(temp_filename1); //deleting the temp file
+
+    exec_view(hunt_id, id);
 }
 
 void start_monitor()
@@ -95,17 +191,27 @@ void start_monitor()
         sa_list_hunts.sa_flags = 0; //using default behavior of sa_flags
         if(sigaction(SIGUSR1, &sa_list_hunts, NULL) == -1) //action to be taken upon the incoming specific signal
         {
-            perror("(start_monitor) sigaction SIGUSR1:");
+            perror("(start_monitor) sigaction SIGUSR1");
             exit(1);
         }
 
         struct sigaction sa_list_treasures;
         sa_list_treasures.sa_handler = list_treasures;
-        sigemptyset(&sa_list_hunts.sa_mask);
-        sa_list_hunts.sa_flags = 0;
+        sigemptyset(&sa_list_treasures.sa_mask);
+        sa_list_treasures.sa_flags = 0;
         if(sigaction(SIGUSR2, &sa_list_treasures, NULL) == -1)
         {
-            perror("(start_monitor) sigaction SIGUSR2:");
+            perror("(start_monitor) sigaction SIGUSR2");
+            exit(1);
+        }
+
+        struct sigaction sa_view_treasure;
+        sa_view_treasure.sa_handler = view_treasure;
+        sigemptyset(&sa_view_treasure.sa_mask);
+        sa_view_treasure.sa_flags = 0;
+        if(sigaction(SIGPIPE, &sa_view_treasure, NULL) == -1)
+        {
+            perror("(start_monitor) sigaction SIGPIPE");
             exit(1);
         }
 
@@ -154,11 +260,27 @@ int main(void)
                 printf("Monitor not running. Please start it first!\n");
             else
             {
+                char cmd_hunt_id[MAX_HUNT_ID_LEN];
                 printf("Sending command signal to monitor...\n");
                 printf("Please enter a hunt ID:\n");
-                fgets(cmd_hunt_id, MAX_HUNT_ID_LEN, stdin);
                 getchar();
+                fgets(cmd_hunt_id, MAX_HUNT_ID_LEN, stdin);
                 cmd_hunt_id[strcspn(cmd_hunt_id, "\n")] = '\0';
+
+                char temp_filename[64];
+                strcat(temp_filename, "/tmp/list_treasures_");
+                snprintf(temp_filename, sizeof(temp_filename), "list_treasures_%d", getpid()); //using the current process pid for a unique file name
+                int fd = open(temp_filename, O_WRONLY | O_CREAT | O_TRUNC, 0600);
+                if(fd == -1)
+                {
+                    perror("(main) open temp:");
+                    exit(1);
+                }
+                //printf("STRING: %s\n", cmd_hunt_id);
+                write(fd, cmd_hunt_id, strlen(cmd_hunt_id));
+                close(fd);
+
+
                 if(kill(monitor_pid, SIGUSR2) == -1)
                     perror("(main) kill SIGUSR2:");
             }
@@ -170,6 +292,41 @@ int main(void)
             else
             {
                 printf("Sending command signal to monitor..\n");
+
+                char hunt_id[MAX_HUNT_ID_LEN];
+                char id[MAX_TREASURE_ID_LEN];
+                printf("Please enter a hunt ID:\n");
+                getchar();
+                fgets(hunt_id, MAX_HUNT_ID_LEN, stdin);
+                hunt_id[strcspn(hunt_id, "\n")] = '\0';
+                printf("Please enter treasure ID:\n");
+                fgets(id, MAX_TREASURE_ID_LEN, stdin);
+                id[strcspn(id, "\n")] = '\0';
+
+                char temp_filename[64];
+                snprintf(temp_filename, sizeof(temp_filename), "view_treasure_hunt_id_%d", getpid());
+                int fd = open(temp_filename, O_WRONLY | O_CREAT | O_TRUNC, 0600);
+                if (fd == -1)
+                {
+                    perror("(main) open");
+                    exit(1);
+                }
+                write(fd, hunt_id, strlen(hunt_id));
+                close(fd);
+
+                char temp_filename1[64];
+                snprintf(temp_filename1, sizeof(temp_filename1), "view_treasure_id_%d", getpid());
+                int fd1 = open(temp_filename1, O_WRONLY | O_CREAT | O_TRUNC, 0600);
+                if (fd1 == -1)
+                {
+                    perror("(main) open");
+                    exit(1);
+                }
+                write(fd1, id, strlen(id));
+                close(fd1);
+
+                if (kill(monitor_pid, SIGPIPE) == -1)
+                    perror("(main) kill SIGPIPE");
             }
         }
         else if(cmd == 5)
